requirements.txt
gdspy
pillow